refactor: Separate UI from business rule
